# internal/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: EventPulse API
  version: "1.0.0"
  description: |
    EventPulse is a high-throughput event ingestion and analytics backend.
    Client applications POST events to the ingestion API, which validates,
    rate-limits, and enqueues them. A worker service persists events to
    PostgreSQL and updates aggregate counters. Analytics endpoints surface
    the stored data.

    **Authentication**: Send `Authorization: Bearer <api_key>` on all `/v1` routes.
    API keys are prefixed with `epk_`.
servers:
  - url: https://ingestion-api-production-137c.up.railway.app
    description: Production (Railway)
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "epk_<hex>"
      description: API key issued per project. Prefix must be `epk_`.
  schemas:
    ValidationError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: event
        message:
          type: string
          example: "event name is required"
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable description
          example: "validation failed"
        details:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
...
    EventRow:
      type: object
      required: [id, event, timestamp, received_at]
      properties:
        id:
          type: string
          format: uuid
        event:
          type: string
          example: page_viewed
        user_id:
          type: string
          example: user_42
        properties:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        received_at:
          type: string
          format: date-time
...
    EventsListResponse:
      type: object
      required: [events, total, limit, offset]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventRow"
        total:
          type: integer
          description: Total matching events (before pagination)
          example: 5000
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
...
paths:
...
  /v1/projects/{projectID}/events:
    get:
      tags: [analytics]
      summary: List events for a project
      description: |
        Returns a paginated list of raw events for the project, newest first.
        Supports filtering by event name, user ID, and time range.
      operationId: listEvents
      security:
        - bearerAuth: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 1000
          description: Maximum number of events to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
        - name: event
          in: query
          schema:
            type: string
          description: Filter by exact event name
          example: page_viewed
        - name: user_id
          in: query
          schema:
            type: string
          description: Filter by user ID
          example: user_42
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Return events at or after this ISO 8601 timestamp
          example: "2024-06-01T00:00:00Z"
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: Return events before this ISO 8601 timestamp
      responses:
        "200":
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsListResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...